services/medical-records-service/requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
prometheus-fastapi-instrumentator==7.0.0
opentelemetry-api==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-exporter-otlp==1.24.0
pydantic==2.7.1
pytest==8.2.2
motor
pymongo
